backend/requirements.txt

fastapi
uvicorn
torch
torchvision
python-multipart
pillow
